+++ internal/protocol/flex.go
@@ -58,6 +58,10 @@ func flexibleRequestHeader(apiKey, apiVersion int16) bool {
 		return apiVersion >= 2
 	case APIElectLeaders:
 		return apiVersion >= 2
+	case APIDescribeLogDirs:
+		return apiVersion >= 2
+	case APIAlterUserScramCreds:
+		return apiVersion >= 0
 	case APIIncrementalAlterConfigs:
 		return apiVersion >= 1
 	case APICreateAcls, APIDescribeAcls, APIDeleteAcls: